+++ src/library/CMakeLists.txt
@@ -12,7 +12,7 @@ add_library(${PROJECT_NAME} STATIC
   logger/logger.cpp
   logger/initialize_log.cpp
 
-  gravity_potential/gravity_potential.cpp
+  gravity/gravity_potential.cpp
 
   randomization/global_randomization.cpp
   randomization/normal_randomization.cpp
