+++ src/library/gravity/gravity_potential.cpp
